Added a file to deal with SED excesses.
This file will pretty much be the location of functions which handle looking
for SED excesses in binaries. Much of this arises from the Casagrande et al
(2010) paper. This includes both exploratory code to find which bands are the
most useful as well as code to actually detect color excesses in actual data.