Working on reproducing the McQuillan sample.
Part of this is working with the original KIC. The full KIC crashes due to lack
of memory. So I have to work with the KIC on CasJobs. As a result, I have
functions to move data in and out of CasJobs. I managed to get the targets with
Kepler observations in. The next thing I need are the Davis and Charbeanneau
properties.