Managed to successfully make the pdf and cdf plots.
However, there are a few major things that need to be investigated:

1) The Vsini > 30 km/s region should have only 4/77=5% of the power. But it
instead has 10% of the power. Look into why this is the case (better numerical
resolution may be the solution)

  a) One thing to look at is the individual distribution functions in that
region. Should they all add up to 10%?

2) When numerical resolution is increased enough (let's say 400), the error in
the integral goes up dramatically! Why is this the case? Am I not multiplying
by dv in the correct place?