Added V-band magnitudes to observing target selection.
This take V-band magnitudes from the EHK catalog. In order to improve
performance, I've adjusted the catalog read functions to only read in the
overlap sample with McQuillan. That way, only relevant segments of the catalog
will have to be stored in memory.

Additionally, it seems like the memoized functions are overwritten each time
the "catalog" module is reloaded. In this case, I think it would be useful to
transfer all of the functions directly dealing with reading data in and out to
a different module: one that will not be reloaded as often.