Integrated *exact* sini distributions into the code.
This should lead to strong performance enhancements because many vsinis no
longer need to be binned. They just need to be evaluated analytically. However,
the fact that we must simultaneously convolve the sin(i) distribution with the
error means that the full distibution can't be computed analytically. However,
if stated in the right way, it should be able to be calculated numerically to a
given precision. This will lead to being able to truly use a K-S test instead
of the chi-squared test. I'm not sure if this will be worth the work, but it's
something to think about, and just to show that I can do.