Found an exact treatment for the vsini distribution of a single object.
This should improve performance of this function since a large number of random
sinis no longer have to be manipulated. This also results in an exact number
without having to montecarlo anything. While the vsin(i) of all the objects can
now be written as an analytic function, things are more complicated because of
the uncertainties. Unfortunately, the Gaussian errors of the uncertainties
cause the full distribution to no longer be analytic. However, we successfully
have approximated it by binning.

I may try to look for a way incorporate the uncertainties such that they can be
determined numerically without binning. This will allow a true K-S test to be
run, as opposed to comparing binned distributions. However, they should be
equivalent for small enough bins.