Moved many functions from catalog.py to read_catalog.py.
The heavy development of catalog.py meant that memoization was not particularly
useful. Every time catalog.py was reloaded, all of the caches would be purged.
So all functionality having to do with reading in catalogs will be in
read_catalog.py, while functionality with manipulating catalogs will be in
catalog.py.