Added a plot comparing the Bruntt et al (2013) vsinis to APOGEE DR14.
The Bruntt et al (2013) vsinis for Kepler asteroseismic targets provide a
method of evaluating both the detection limits of APOGEE targets and the
uncertainty in the measurements. These results confirm the initial parameters
of a vsini upper limit at 7 km/s and a vsini uncertainty of ~10%.

I also included a framework to improve the performance of manipulating tables
with a decorator that caches certain tables on-disk so that they can be
recovered without having to be generated all the time. This will be best served
for tasks which require a significant amount of table processing that ends up
culling a table significantly, so the end product is not large.