Modified find_DLSBs to only use IDs and added whitespace stripping.
In order to make better use of memory, I removed table manipulation from the
function. It only requires the apogee_ids and loc_ids for the targets. I can't
be sure, but this seems to be faster than the previous way.

Additionally, APOGEE_IDs, both provided and in the DLSB database are stripped
when they are compared. Too often this leads to false inequalities, causing
duplication of labor.