Compare Huber evolutionary state classification to APOGEE.
Plot to show how well the evolutionary state classifications in Huber agree
with those in APOGEE. According to this figure, around half of the Huber
subgiants are actually dwarfs, meaning that the radii are overestimated. This
exacerbates the issue with vsinis being larger than predicted because smaller
radii mean smaller predicted velocities.

Oddly enough, there doesn't seem to be a bias in terms of which objects have
detected vsinis either. Indicating that the vsini problem seems unrelated to
an object's location on the HR diagram. This indicates that the issue is
unrelated to radii measurements.

It may be useful to check out the hot stars to see if non-asteroseismic targets
experience the same problem.