Trying to get to the point where I can cross-match the rapid rotators.
Right now I am able to identify the spectroscopic rapid rotators in the DR14
dataset. I'd like to see if those spectroscopic rapid rotators correspond to
photometric rapid rotators in McQuillan, so the two datasets will have to
harmoniously interact. Currently in the middle of making functions that will do
that.