Interface for 2MASS and generate distance modulii.
The two main changes in this commit are the creation of functions to download
entries from 2MASS. TAP ended up being a failure and I just created an IPAC
file to upload to the IRSA web site.

All of the Gaia conversion routines have now moved to directly converting
parallaxes to distance modulii.